rfp-extraction/pyproject.toml
[project]
name = "rfp-extraction"
version = "0.1.0"
description = "PDF extraction Azure Function for RFP Management System"
requires-python = ">=3.10"
dependencies = [
    # Azure Functions core
    "azure-functions>=1.24.0",
    "azure-identity>=1.25.1",
    "azure-storage-blob>=12.27.1",

    # Azure AI services
    "azure-ai-documentintelligence>=1.0.0b1",
    "azure-ai-formrecognizer>=3.3.3",

    # OpenAI
    "openai>=1.58.1",

    # Database
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.2.0",

    # Data validation
    "pydantic>=2.10.4",
    "pydantic-settings>=2.0.0",

    # CLI and utilities
    "watchdog>=3.0.0",  # For directory watching
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",

    # Code quality
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]